markdown-preview: fix errors from build
Not sure what underlying issue was, but the repo
contained 2 methods of configuring this plugin, one
that requires yarn and node locally and one that
doesn't. This commit switches to using the method that
does not require yarn and node locally. This appears to
have resolved the build errors for now.