refactor(neorg): enhance autocmd setup and buffer local settings
This commit refactors the autocmd setup for Neorg files and enhances
buffer local settings. The autocmd now triggers on `BufNew` event
instead of `BufEnter` and `BufWinEnter`, which makes it more efficient
as it only triggers for new buffers.

The buffer local settings have been enhanced to set `conceallevel`,
`nowrap`, `textwidth`, and `spell` only for the current buffer. This
ensures that these settings do not affect other buffers.

The commit also includes changes to the which-key bindings for Neorg.
The bindings are now set for the current buffer only, and a new group
for Todo related commands has been added.